requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.19.1
rich==13.9.4
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
